fastapi/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
httpx==0.28.1
loguru==0.7.3
python-dotenv==1.0.1
python-docx==1.2.0
